requirements.txt
beautifulsoup4
requests
requests-html
lxml_html_clean
selenium
pyautogen>=0.2.3